MAJOR CHANGE TO FUNCTIONALITY
This script will now separate tables if there is too large a time gap
between two entries, resulting in multiple smaller tables/graphs.
This change is to allow for a single log file to not result in a very visually
confusing graph if that log file is not being continuously updated.

IMPORTANT: the --outputfile argument HAS BEEN DEPRECATED. This is because multiple
output files are now potentially being created and are instead stored in the new Output
directory.

TODO: allow for dynamic naming of the Output directory.